+++ CHANGES.md
@@ -1,3 +1,5 @@
+* New: Banner alignment setting.
+
 #### 1.0.2.1
 * Hotfix: z-index issue on sections template prohibiting editing of some fields.